requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# API client
requests>=2.31.0
entsoe-py>=0.6.0

# Configuration
python-dotenv>=1.0.0

# Visualization (for future dashboard)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Web dashboard (Week 4)
streamlit>=1.29.0

# Statistical modeling (Week 3)
statsmodels>=0.14.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Development tools
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.27.0

# Nordic Power Price Dashboard - Requirements
# Week 1-3: Data Pipeline + GARCH Forecasting

# Core dependencies (Week 1-2)
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
entsoe-py>=0.6.8
python-dotenv>=1.0.0

# GARCH modeling (Week 3)
arch>=5.3.0              # GARCH/ARCH models
statsmodels>=0.14.0      # Statistical tests and models
scipy>=1.11.0            # Scientific computing
scikit-learn>=1.3.0      # ML utilities (Linear Regression for MZ test)

# Visualization (Week 3)
matplotlib>=3.7.0        # Plotting
seaborn>=0.12.0          # Statistical visualization

# Optional for Week 4 (Streamlit dashboard)
# streamlit>=1.28.0
# plotly>=5.17.0
